Refactor get_sensitivity_access into find_flashes
However, this now has a custom Exception type that will be raised if the
SOAP API raises a Share Fault.

ShareErrors are currently a bit general; this happens for a number of
reasons, if the person does not exist, if we can't access the person, or
if the ID is malformed. Right now, we just yell "RED FLAG" at the call
site, but we should likely do something smarter.

However, this likely means the BGS team needs to add a thing. So, for
now, this.